requirements.txt
# Core dependencies
pyspark==3.5.0
kafka-python==2.0.2
pandas==2.1.4
numpy==1.24.3

# ML
scikit-learn==1.3.2

# Cloud SDKs
google-cloud-storage==2.14.0
google-cloud-aiplatform==1.38.1
google-cloud-functions==1.13.0
boto3==1.34.0

# Monitoring
flask==3.0.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0

